# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Auto Forex Trading API
  version: 1.0.0
  description:
    API for managing algorithmic forex trading operations across multiple
    OANDA accounts
  contact:
    name: Auto Forex Support
    email: support@autoforex.example.com
  license:
    name: Proprietary
paths:
...
  /api/accounts/auth/logout:
    post:
      operationId: accounts_auth_logout_create
      description:
        Logout user and terminate all active sessions. Requires valid JWT
        token in Authorization header.
      summary: POST /api/accounts/auth/logout
      tags:
        - Authentication
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  sessions_terminated:
                    type: integer
          description: Logout successful
        "401":
          description: Invalid or expired token
...
  /api/accounts/notifications:
    get:
      operationId: accounts_notifications_retrieve
      description:
        Retrieve list of notifications for the authenticated user. Supports
        filtering by read status and pagination.
      summary: GET /api/accounts/notifications
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: page_size
          schema:
            type: integer
          description: "Number of results per page (default: 50, max: 200)"
        - in: query
          name: unread_only
          schema:
            type: boolean
          description: Filter to show only unread notifications
      tags:
        - Notifications
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Paginated notifications list
        "401":
          description: Authentication required
        "500":
          description: Failed to retrieve notifications
  /api/accounts/notifications/{notification_id}/read:
    post:
      operationId: accounts_notifications_read_create
      description: Mark a single notification as read for the authenticated user.
      summary: POST /api/accounts/notifications/{notification_id}/read
      parameters:
        - in: path
          name: notification_id
          schema:
            type: integer
          description: ID of the notification to mark as read
          required: true
      tags:
        - Notifications
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Notification marked as read
        "401":
          description: Authentication required
        "404":
          description: Notification not found
        "500":
          description: Failed to mark notification as read
  /api/accounts/notifications/read-all:
    post:
      operationId: accounts_notifications_read_all_create
      description: Mark all unread notifications as read for the authenticated user.
      summary: POST /api/accounts/notifications/read-all
      tags:
        - Notifications
      security:
        - jwtAuth: []
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  count:
                    type: integer
          description: All notifications marked as read
        "401":
          description: Authentication required
        "500":
          description: Failed to mark all notifications as read
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.autoforex.example.com
    description: Production server
...